move & update tests for set_wp_cache_define
