+++ tests/Unit/inc/Engine/Cache/WPCache/setWpCacheDefine.php
@@ -1,23 +1,20 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\functions;
+namespace WP_Rocket\Tests\Unit\inc\Engine\Cache\WPCache;
 
 use Brain\Monkey\Functions;
+use WP_Rocket\Engine\Cache\WPCache;
 use WP_Rocket\Tests\Unit\FilesystemTestCase;
 
 /**
- * @covers ::set_rocket_wp_cache_define
+ * @covers \WP_Rocket\Engine\Cache\WPCache::set_wp_cache_define
  * @uses   ::rocket_valid_key
- * @uses   ::rocket_find_wpconfig_path
- * @uses   ::rocket_direct_filesystem
- * @uses   ::rocket_put_content
  *
- * @group Functions
- * @group Files
+ * @group WPCache
  * @group vfs
  */
-class Test_SetRocketWpCacheDefine extends FilesystemTestCase {
-	protected $path_to_test_data = '/inc/functions/setRocketWpCacheDefine.php';
+class Test_SetWpCacheDefine extends FilesystemTestCase {
+	protected $path_to_test_data = '/inc/Engine/Cache/WPCache/setWpCacheDefine.php';
 
 	/**
 	 * @dataProvider providerTestData
@@ -28,7 +25,9 @@ class Test_SetRocketWpCacheDefine extends FilesystemTestCase {
 		Functions\when( 'rocket_find_wpconfig_path' )->justReturn( $config_file_full_path );
 		Functions\expect( 'rocket_valid_key' )->once()->andReturn( $config['valid_key'] );
 
-		set_rocket_wp_cache_define( true );
+		$wp_cache = new WPCache( $this->filesystem );
+
+		$wp_cache->set_wp_cache_define( true );
 
 		$actual = $this->filesystem->get_contents( $config_file_full_path );
 		$this->assertEquals( $expected, str_replace( "\r\n", "\n", $actual ) );
